fixed some things

+++ backend_v2/src/main/java/com/cyber/sentinel/backend/model/Containers/CVE.java
@@ -57,7 +57,7 @@ public class CVE {
     }
 
     public float getScore() {
-        return score;
+        return (float) score;
     }
 
     public void setScore(double score) {
